Extracted styles to separate file
