+++ client/app/components/SignIn/SignIn.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
-import { Text, View, StyleSheet, TouchableHighlight } from 'react-native';
+import { Text, View, TouchableHighlight } from 'react-native';
 import { Card, Input, Icon } from 'react-native-elements';
+import { styles } from './styles';
 
 class SignIn extends Component {
   render() {
@@ -69,51 +70,5 @@ class SignIn extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: '#ddd',
-    // alignItems: 'center',
-    justifyContent: 'center',
-    paddingTop: 22,
-    flex: 1,
-    flexDirection: 'column'
-  },
-  icons: {
-    flexDirection: 'row',
-    justifyContent: 'center'
-  },
-  button: {
-    marginHorizontal: 40,
-    marginTop: 10,
-    paddingVertical: 10,
-    backgroundColor: 'grey',
-    borderRadius: 20,
-    borderWidth: 1,
-    borderColor: '#fff'
-  },
-  buttonText: {
-    color: '#fff',
-    textAlign: 'center',
-    fontSize: 20
-  },
-  divider: {
-    backgroundColor: 'grey',
-    height: 1,
-    flex: 1,
-    alignSelf: 'center'
-  },
-  dividerText: {
-    alignSelf: 'center',
-    paddingHorizontal: 5,
-    paddingVertical: 10,
-    fontSize: 20,
-    color: 'grey'
-  },
-  title: {
-    alignSelf: 'center',
-    fontSize: 40
-  }
-});
 
 export default SignIn;
